restructuring the backend api routes to include a controller
