+++ .gitignore
@@ -14,6 +14,7 @@ pids
 *.pid
 *.seed
 *.pid.lock
+node_modules
 
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
