+++ models/user.js
@@ -20,10 +20,12 @@ const userSchema = new Schema({
       unique: true,
       match: [/.+@.+\..+/, "Please enter a valid e-mail address"]
    },
-   userCreated: {
-      type: Date,
-      default: Date.now
-   }
+   notes: [
+      {
+        type: Schema.Types.ObjectId,
+        ref: "Note"
+      }
+    ]
 })
 // Define schema methods
 userSchema.methods = {
@@ -45,5 +47,5 @@ userSchema.pre('save', function (next) {
       next()
    }
 })
-const User = mongoose.model('User', userSchema)
-module.exports = User
+const User = mongoose.model('User', userSchema);
+module.exports = User;
