+++ package.json
@@ -27,6 +27,7 @@
     "if-env": "^1.0.4",
     "mongoose": "^5.9.7",
     "passport": "^0.4.1",
-    "passport-local": "^1.0.0"
+    "passport-local": "^1.0.0",
+    "react-router-dom": "^5.1.2"
   }
 }
