+++ public/index.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en-us">
+  <head>
+    <meta charset="UTF-8">
+    <title>Populate Exercise</title>
+    <style>
+      input {
+        display: block;
+        height: 36px;
+        width: 200px;
+        font-size: 24px;
+      }
+      textarea {
+        width: 200px;
+        height: 200px;
+      }
+    </style>
+  </head>
+  <body>
+    <form action="/signup" method="create">
+      <h1>Hi User!</h1>
+      <h2>Sign Up</h2>
+      <input type="text" name="Username" placeholder="Username">
+      <input type="text" name="Email" placeholder="Email">
+      <input type="text" name="Password" placeholder="Password">
+
+      <input type="submit">
+      <a href="/notes">See your notes!</a>
+      <a href="/user">See your profile!</a>
+      <a href="/populateduser">See your profile, populated with notes!</a>
+    </form>
+  </body>
+</html>
