+++ routes/index.js
@@ -1,0 +1,8 @@
+const path = require("path");
+const router = require("express").Router();
+const apiRoutes = require("./api");
+
+// API Routes
+router.use("/api", apiRoutes);
+
+module.exports = router;
