+++ server.js
@@ -29,7 +29,7 @@ app.use(passport.initialize())
 app.use(passport.session()) // calls the deserializeUser
 
 // Define API routes here
-const userRoutes = require("./API_routes/user-route");
+const userRoutes = require("./routes/api/user-route");
 app.use("/user", userRoutes);
 // Send every other request to the React app
 // Define any API routes before this runs
